Extract flag computation into shared util
